+++ highcharts_core/options/series/treegraph.py
@@ -91,6 +91,7 @@ class TreegraphSeries(SeriesBase, TreegraphOptions):
             'show_checkbox': as_dict.get('showCheckbox', None),
             'show_in_legend': as_dict.get('showInLegend', None),
             'skip_keyboard_navigation': as_dict.get('skipKeyboardNavigation', None),
+            'sonification': as_dict.get('sonification', None),
             'states': as_dict.get('states', None),
             'sticky_tracking': as_dict.get('stickyTracking', None),
             'tooltip': as_dict.get('tooltip', None),